refactor: improve folders structure for components
