+++ src/Components/AboutUs/AboutUs.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Header } from './Header';
+import { Header } from '../Header/Header';
 
 export class AboutUs extends React.Component {
   render() {
