+++ src/Components/ErrorNotFound/ErrorNotFound.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Header } from './Header';
+import { Header } from '../Header/Header';
 
 export class ErrorNotFound extends React.Component {
   render() {
